+++ CHANGELOG.md
@@ -71,6 +71,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Changes look of landing page and About modal buttons
 - Changes size and transparency of unflagged points to help identify overlapping points
 - Put units on second line in graph axis titles
+- Changes autotick to true for log scale
 
 ### Fixed